adjust failing provider tests
