+++ tests/Hateoas/Tests/Configuration/Provider/ExpressionEvaluatorProviderTest.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace Hateoas\Tests\Configuration\Provider;
+
+use Hateoas\Configuration\Provider\ExpressionEvaluatorProvider;
+use Hateoas\Configuration\RelationProvider as RelationProviderConfiguration;
+use Hateoas\Tests\TestCase;
+use JMS\Serializer\Expression\ExpressionEvaluatorInterface;
+
+class ExpressionEvaluatorProviderTest extends TestCase
+{
+    public function test()
+    {
+        $service = new \StdClass();
+
+        $containerProphecy = $this->prophesize(ExpressionEvaluatorInterface::class);
+        $containerProphecy
+            ->evaluate('foo', ['class' => \stdClass::class])
+            ->willReturn([$service])
+        ;
+
+        $providerProvider = new ExpressionEvaluatorProvider($containerProphecy->reveal());
+
+        $this->assertEmpty($providerProvider->getRelations(new RelationProviderConfiguration('!-;'), \stdClass::class));
+        $this->assertEmpty($providerProvider->getRelations(new RelationProviderConfiguration('getSomething'), \stdClass::class));
+        $this->assertEmpty($providerProvider->getRelations(new RelationProviderConfiguration('foo::bar'), \stdClass::class));
+        $this->assertSame(
+            [$service],
+            $providerProvider->getRelations(new RelationProviderConfiguration('expr(foo)'), \stdClass::class)
+        );
+    }
+}
