+++ tests/Hateoas/Tests/Configuration/Provider/StaticMethodProviderTest.php
